+++ README.md
@@ -18,6 +18,30 @@
 #### `npm run lint`
 > Lints and fixes files - Works for development environment, and uses Eslint to code with Vue JS version 2 standard good practices and rules.
 
+## Folder Structure
+
+In the phase 2 we have simplify the folder structure, you may see this change in some folders and you can adapt old folders to the new structure
+
+- Resources: Common Code
+  - components
+    - form: Inputs
+    - resources: Modals, buttons,
+  - mixins
+  - rules
+  - store: Common stores that can be included in module stores
+
+- Services: All the API services are here, not in the modules.
+
+- Modules:
+  - module: Is the root  component for the router
+  - router: defines the router
+  - component-names: A folder per component
+    - index / [component-name] : view page
+    - store: Component specific store
+    - components: Folder for sub components used only in this page
+    
+
+
 ## Default User
 ### Default Super Admin
