+++ package.json
@@ -3,7 +3,7 @@
   "version": "1.5.3",
   "private": true,
   "scripts": {
-    "serve": "vue-cli-service serve  --open --host localhost",
+    "serve": "vue-cli-service serve",
     "build": "vue-cli-service build",
     "lint": "vue-cli-service lint",
     "dev": "npm run serve"
