--- src/modules/groups/components/CreateOrEditGroups/component/components.js
@@ -1,8 +1,0 @@
-export default {
-  components: {
-    NombreFieldInput: () =>
-      import(
-        /* webpackChunkName: "NombreFieldInput" */ '@/modules/groups/components/form/NombreFieldInput'
-      )
-  }
-}
