--- src/modules/groups/components/CreateOrEditGroups/component/watch.js
@@ -1,13 +1,0 @@
-export default {
-  watch: {
-    currentUserForUpdate (value) {
-      if (value === null) {
-        this.ResetForm()
-
-        return
-      }
-
-      this.loadDataUserForUpdate(value)
-    }
-  }
-}
