+++ src/modules/groups/group-list/index.vue
@@ -1,0 +1,40 @@
+<template>
+  <v-card-text>
+    <div ref="sectionTabsItemsGroupsByStatusAccount">
+      <v-card flat>
+        <v-card-text>
+          <GroupList />
+        </v-card-text>
+      </v-card>
+    </div>
+  </v-card-text>
+</template>
+
+<script>
+import notifications from '@/mixins/notifications'
+
+export default {
+  name: 'GroupsView',
+  components: {
+    GroupList: () =>
+      import(/* webpackChunkName: "GroupList" */ './components/groups-list.vue')
+  },
+  mixins: [notifications],
+  data() {
+    return {
+      reloadDatatableUsers: false
+    }
+  },
+  beforeCreate() {
+    this?.$hasRoleMiddleware('admin')
+  },
+  mounted() {
+    this.loadNotifications()
+  },
+  head: {
+    title: {
+      inner: 'Gestión de grupos'
+    }
+  }
+}
+</script>
