+++ src/modules/groups/group-students/index.vue
@@ -1,0 +1,42 @@
+<template>
+  <v-card-text>
+    <div ref="sectionTabsItemsGroupsByStatusAccount">
+      <v-card flat>
+        <v-card-text>
+          <students-datatable />
+        </v-card-text>
+      </v-card>
+    </div>
+  </v-card-text>
+</template>
+
+<script>
+import notifications from '@/mixins/notifications'
+
+export default {
+  name: 'StudentsView',
+  components: {
+    StudentsDatatable: () =>
+      import(
+        /* webpackChunkName: "DatatableStudents" */ './components/groups-students-list.vue'
+      )
+  },
+  mixins: [notifications],
+  data() {
+    return {
+      reloadDatatableUsers: false
+    }
+  },
+  beforeCreate() {
+    this?.$hasRoleMiddleware('admin')
+  },
+  mounted() {
+    this.loadNotifications()
+  },
+  head: {
+    title: {
+      inner: 'Gestión de grupos'
+    }
+  }
+}
+</script>
