+++ src/modules/groups/index.js
@@ -1,7 +1,9 @@
-import store from './group.store'
+import groupStore from './group-list/group.store'
+import groupStudentsStore from './group-students/group-students.store'
+
 import router from './group.router'
 
 export default {
-  store,
+  stores: [groupStore, groupStudentsStore],
   router
 }
