+++ src/modules/students/store/actions.js
@@ -1,4 +1,4 @@
-import StudentRepository from '../repositories/StudentRepository'
+import StudentRepository from '../../../services/StudentRepository'
 
 const mapItemsDatatableFromApi = (itemsApi) => {
   return itemsApi.map((record) => {
@@ -26,7 +26,6 @@ const mapMetaInformationPagination = (response) => {
 }
 
 const getStudents = async ({ commit }, config) => {
-
   try {
     commit('SET_USERS_SELECTED_DATATABLE', [])
     commit('SET_INFORMATION_META', {
@@ -46,14 +45,15 @@ const getStudents = async ({ commit }, config) => {
     if (response) {
       //console.trace(response)
 
-      commit('SET_ITEMS_DATATABLE', mapItemsDatatableFromApi(response.data.data))
+      commit(
+        'SET_ITEMS_DATATABLE',
+        mapItemsDatatableFromApi(response.data.data)
+      )
       commit('SET_STATUS_LOADING_ITEMS', false)
       commit('SET_INFORMATION_META', mapMetaInformationPagination(response))
-
     }
 
     return Promise.resolve(response)
-
   } catch (error) {
     //console.log(error)
 
@@ -74,7 +74,6 @@ const createStudent = async (_, options) => {
 }
 
 const fetchStudent = async ({ commit }, options) => {
-
   try {
     commit('SET_USERS_SELECTED_DATATABLE', [])
     const response = await StudentRepository.get(options.id, options.config)
@@ -89,7 +88,11 @@ const fetchStudent = async ({ commit }, options) => {
 
 const updateStudent = async (_, options) => {
   try {
-    const response = await StudentRepository.update(options.id, options.data, options.config)
+    const response = await StudentRepository.update(
+      options.id,
+      options.data,
+      options.config
+    )
 
     return Promise.resolve(response)
   } catch (error) {
@@ -125,7 +128,9 @@ const fetchStudentGroups = async (_, options) => {
 
 const fetchRoleStudentData = async (_, options) => {
   try {
-    const response = await StudentRepository.fetchRoleStudentData(options.config)
+    const response = await StudentRepository.fetchRoleStudentData(
+      options.config
+    )
 
     return Promise.resolve(response)
   } catch (error) {
@@ -137,7 +142,10 @@ const fetchRoleStudentData = async (_, options) => {
 
 const actionsForMultipleRecords = async ({ commit }, options) => {
   try {
-    const response = await StudentRepository.actionMassiveSelection(options.data, options.config)
+    const response = await StudentRepository.actionMassiveSelection(
+      options.data,
+      options.config
+    )
 
     commit('SET_USERS_SELECTED_DATATABLE', [])
